Updates for moss-client: 1. Force service discovery to ZooKeeper for moss-client both 2.x and 1.x. 2. Remove '/infoLogFile' and '/errorLogFile' endpoint, use '/logfile' and '/logfile/{requiredLogFileName}'. 3. Add logFileRegistry for '/logfile' endpoint. 4. Change '/appInfo' to '/appinfo' endpoint. 5. Add hateoas support for moss-client 1.x. 6. Add '/prometheus' endpoint for both client by default. 7. Auto collect metadata for registration in {AutoRegistrationCustomizer} and {MetaDataProvider} 8. Support on outer servlet container such as Tomcat.    @see {DiscoveryClientRegistrationInvoker}    DiscoveryClient registration depends on {EmbeddedServletContainerInitializedEvent} in 1.5.x    and {ServletWebServerInitializedEvent} in Spring Boot 2.x. The registration logic is in    {AbstractAutoServiceRegistration} in Spring Boot 2.x and {AbstractDiscoveryLifecycle}.    The spring cloud official has't been fixed it on thier repository issues, guess that won't    be fixed.    @see    (https://github.com/spring-cloud/spring-cloud-commons/commit/1d646290abb5a123e452225178a355593f85565a).    So we had to manually publish an event for registration in outer Tomcat. 9. Add some default bootstrap configuration in META-INF/moss-client/bootstrap.properties.    It will loaded by {ManagementEnvironmentCustomizer}. And the environment customizer    is registered by spring.factories.
Add the new client
