+++ moss-client/moss-client-1.x/src/main/java/org/xujin/moss/client/config/AdminEndPointConfiguration.java
@@ -2,9 +2,9 @@ package org.xujin.moss.client.config;
 
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.x.discovery.details.InstanceSerializer;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringBootVersion;
-import org.springframework.boot.actuate.info.SimpleInfoContributor;
+import org.springframework.boot.actuate.endpoint.PublicMetrics;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cache.CacheManager;
@@ -13,8 +13,14 @@ import org.springframework.cloud.zookeeper.discovery.ZookeeperInstance;
 import org.springframework.cloud.zookeeper.support.ServiceDiscoveryCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.AnnotationAwareOrderComparator;
 import org.springframework.core.env.Environment;
 import org.xujin.moss.client.endpoint.*;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 
 @Configuration
 @EnableConfigurationProperties(LogFileRegistry.class)
@@ -22,6 +28,14 @@ public class AdminEndPointConfiguration {
 
     @Autowired
     private Environment env;
+
+
+    private final Collection<PublicMetrics> publicMetrics;
+
+
+    public AdminEndPointConfiguration(ObjectProvider<Collection<PublicMetrics>> publicMetrics) {
+        this.publicMetrics = publicMetrics.getIfAvailable();
+    }
 
     @Bean
     public JarDependenciesEndpoint jarDependenciesEndpoint() {
@@ -41,7 +55,12 @@ public class AdminEndPointConfiguration {
 
     @Bean
     public MossMetricsEndpoint haloMetricsEndpoint(){
-        return new MossMetricsEndpoint();
+        List<PublicMetrics> publicMetrics = new ArrayList<PublicMetrics>();
+        if (this.publicMetrics != null) {
+            publicMetrics.addAll(this.publicMetrics);
+        }
+        Collections.sort(publicMetrics, AnnotationAwareOrderComparator.INSTANCE);
+        return new MossMetricsEndpoint(publicMetrics);
     }
 
     @Bean
@@ -55,17 +74,13 @@ public class AdminEndPointConfiguration {
         return new CacheManagerEndpoint();
     }
 
-    @Bean
-    public SimpleInfoContributor springBootVersionInfoContributor() {
-        return new SimpleInfoContributor("spring-boot-version", SpringBootVersion.getVersion());
-    }
+
     @Bean
     public ServiceDiscoveryCustomizer defaultServiceDiscoveryCustomizer(CuratorFramework curator,
-                                                                               ZookeeperDiscoveryProperties properties,
-                                                                               InstanceSerializer<ZookeeperInstance> serializer) {
+                                                                        ZookeeperDiscoveryProperties properties,
+                                                                        InstanceSerializer<ZookeeperInstance> serializer) {
         return new AutoRegistrationCustomizer(curator, properties, serializer);
     }
-
     @Bean
     public MetaDataProvider metaDataProvider() {
         return new MetaDataProvider();
