+++ moss-client/moss-client-1.x/src/main/java/org/xujin/moss/client/config/AdminEndpointApplicationRunListener.java
@@ -16,46 +16,34 @@ import java.util.List;
 
 public class AdminEndpointApplicationRunListener implements SpringApplicationRunListener {
 
-
-
-    public AdminEndpointApplicationRunListener(SpringApplication application, String[] args) {
-    }
-
     @Override
-    public void contextLoaded(ConfigurableApplicationContext context) {
-    }
+    public void starting() {
 
-    @Override
-    public void contextPrepared(ConfigurableApplicationContext context) {
     }
 
     @Override
     public void environmentPrepared(ConfigurableEnvironment env) {
         if("bootstrap".equals(env.getProperty("spring.config.name"))) {
             List<EnvironmentCustomizer> environmentCustomizers =
-            SpringFactoriesLoader.loadFactories(EnvironmentCustomizer.class, AdminEndpointApplicationRunListener.class.getClassLoader());
+                    SpringFactoriesLoader.loadFactories(EnvironmentCustomizer.class, AdminEndpointApplicationRunListener.class.getClassLoader());
             if(CollectionUtils.isEmpty(environmentCustomizers)) return;
             for(EnvironmentCustomizer customizer: environmentCustomizers) {
                 customizer.customize(env);
             }
         }
+    }
+    @Override
+    public void contextPrepared(ConfigurableApplicationContext configurableApplicationContext) {
 
     }
 
     @Override
-    public void failed(ConfigurableApplicationContext context, Throwable exception) {
-
-    }
-
-
-    @Override
-    public void started(ConfigurableApplicationContext context) {
+    public void contextLoaded(ConfigurableApplicationContext configurableApplicationContext) {
 
     }
 
     @Override
-    public void running(ConfigurableApplicationContext context) {
-
+    public void finished(ConfigurableApplicationContext context, Throwable throwable) {
         Collection<ApplicationContextCustomizer> applicationContextCustomizers =
                 Arrays.asList(new DiscoveryClientRegistrationInvoker());
         for(ApplicationContextCustomizer customizer: applicationContextCustomizers) {
@@ -63,8 +51,7 @@ public class AdminEndpointApplicationRunListener implements SpringApplicationRun
         }
     }
 
-    @Override
-    public void starting() {
+    public AdminEndpointApplicationRunListener(SpringApplication application, String[] args) {
     }
     public static boolean isEmbeddedServletServer(Environment env) {
         return StringUtils.isEmpty(env.getProperty(ContextLoader.CONFIG_LOCATION_PARAM));
