+++ moss-client/moss-client-1.x/src/main/java/org/xujin/moss/client/config/EnvironmentCustomizer.java
@@ -1,0 +1,7 @@
+package org.xujin.moss.client.config;
+
+import org.springframework.core.env.Environment;
+
+public interface EnvironmentCustomizer<T extends Environment> {
+    void customize(T t);
+}
