+++ moss-client/moss-client-1.x/src/main/java/org/xujin/moss/client/config/ManagementConfig.java
@@ -1,0 +1,13 @@
+package org.xujin.moss.client.config;
+
+import org.springframework.boot.actuate.autoconfigure.ManagementContextConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+
+/**
+ * @Author: xujin
+ **/
+@ManagementContextConfiguration
+@ConditionalOnWebApplication
+public class ManagementConfig {
+
+}
