+++ moss-client/moss-client-1.x/src/main/resources/META-INF/moss-client/bootstrap.properties
@@ -1,0 +1,15 @@
+spring.cloud.zookeeper.discovery.root=/moss-client
+management.info.git.mode=full
+management.context-path=/actuator
+management.security.enabled=false
+endpoints.actuator.enabled=true
+endpoints.actuator.sensitive=false
+endpoints.metrics.enabled=true
+endpoints.mappings.enabled=true
+endpoints.restart.enabled=false
+endpoints.shutdown.enabled=false
+endpoints.heapdump.enabled=false
+endpoints.pause.enabled=false
+endpoints.resume.enabled=false
+endpoints.refresh.enabled=false
+endpoints.sensitive=false
