+++ moss-client/moss-client-1.x/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,8 @@
-org.springframework.boot.actuate.autoconfigure.ManagementContextConfiguration=org.xujin.moss.config.AdminEndPointConfiguration
-org.springframework.boot.SpringApplicationRunListener=org.xujin.moss.config.AdminEndpointApplicationRunListener
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    org.xujin.moss.client.config.AdminEndPointConfiguration
+org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration=\
+    org.xujin.moss.client.config.ManagementConfig
+org.springframework.boot.SpringApplicationRunListener=\
+    org.xujin.moss.client.config.AdminEndpointApplicationRunListener
+org.xujin.moss.client.config.EnvironmentCustomizer=\
+    org.xujin.moss.client.config.ManagementEnvironmentCustomizer
