+++ moss-client/moss-client-2.x/src/main/java/org/xujin/moss/client/config/ApplicationContextCustomizer.java
@@ -1,0 +1,7 @@
+package org.xujin.moss.client.config;
+
+import org.springframework.context.ConfigurableApplicationContext;
+
+public interface ApplicationContextCustomizer<T extends ConfigurableApplicationContext> {
+    void customize(T t);
+}
