+++ moss-client/moss-client-2.x/src/main/java/org/xujin/moss/client/config/ManagementConfig.java
@@ -1,10 +1,8 @@
 package org.xujin.moss.client.config;
 
-import org.xujin.moss.client.endpoint.PlatformManagementWebServerFactoryCustomizer;
 import org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration;
 import org.springframework.boot.actuate.autoconfigure.web.ManagementContextType;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.context.annotation.Bean;
 
 /**
  * @Author: xujin
@@ -12,10 +10,5 @@ import org.springframework.context.annotation.Bean;
 @ManagementContextConfiguration(ManagementContextType.ANY)
 @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
 public class ManagementConfig {
-
-    @Bean
-    public PlatformManagementWebServerFactoryCustomizer healthExposableEndpoint(){
-        return new PlatformManagementWebServerFactoryCustomizer();
-    }
 
 }
