+++ moss-client/moss-client-2.x/src/main/resources/META-INF/moss-client/bootstrap.properties
@@ -1,0 +1,23 @@
+spring.cloud.zookeeper.discovery.root=/moss-client
+management.endpoints.web.exposure.include=*
+management.endpoints.web.exposure.exclude[0]=pause
+management.endpoints.web.exposure.exclude[1]=refresh
+management.endpoints.web.exposure.exclude[2]=restart
+management.endpoints.web.exposure.exclude[3]=resume
+management.endpoints.web.exposure.exclude[4]=shutdown
+management.endpoints.web.exposure.exclude[5]=heapdump
+management.endpoints.jmx.exposure.include=*
+management.endpoints.jmx.exposure.exclude[0]=pause
+management.endpoints.jmx.exposure.exclude[1]=refresh
+management.endpoints.jmx.exposure.exclude[2]=restart
+management.endpoints.jmx.exposure.exclude[3]=resume
+management.endpoints.jmx.exposure.exclude[4]=shutdown
+management.endpoints.jmx.exposure.exclude[5]=heapdump
+management.endpoint.health.show-details=always
+management.endpoint.heapdump.enabled=false
+management.endpoint.pause.enabled=false
+management.endpoint.refresh.enabled=false
+management.endpoint.restart.enabled=false
+management.endpoint.resume.enabled=false
+management.endpoint.shutdown.enabled=false
+management.info.git.mode=full
