+++ moss-client/moss-sample-1.5.x/src/main/java/org/xujin/moss/sample/Sb15XAppliaction.java
@@ -1,5 +1,7 @@
 package org.xujin.moss.sample;
 
+import io.prometheus.client.spring.boot.EnablePrometheusEndpoint;
+import io.prometheus.client.spring.boot.EnableSpringBootMetricsCollector;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -10,6 +12,8 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
  * @site http://xujin.org
  */
 @SpringBootApplication
+@EnableSpringBootMetricsCollector
+@EnablePrometheusEndpoint
 @EnableDiscoveryClient
 public class Sb15XAppliaction {
