+++ moss-client/moss-sample-1.5.x/src/main/resources/config/application.yml
@@ -1,0 +1,14 @@
+server:
+  port: 8085
+spring:
+  application:
+    name: @project.artifactId@
+info.groupId: @project.groupId@
+info.artifactId: @project.artifactId@
+info.version: @project.version@
+
+logging.registry.files[0].name: tomcat_stdout
+logging.registry.files[0].description: 标准输出
+logging.registry.files[0].path: ./spring.log
+logging.registry.files[1].name: gclog
+logging.registry.files[1].path: ${user.home}/logs/sys/${spring.application.name}/gc.log
