+++ moss-client/moss-sample-1.5.x/src/main/resources/config/bootstrap.yml
@@ -1,0 +1,6 @@
+spring:
+  application:
+    name: @project.artifactId@
+  cloud:
+    zookeeper:
+      connect-string: localhost:2181
