+++ moss-client/moss-sample-1.5.x/src/main/resources/logback.xml
@@ -1,0 +1,4 @@
+<configuration>
+    <include resource="org/springframework/boot/logging/logback/base.xml"/>
+    <jmxConfigurator/>
+</configuration>
