--- moss-client/moss-sample-2.1.x/src/main/resources/application.yml
@@ -1,19 +1,0 @@
-spring:
-  profiles:
-  application:
-    name: moss-sample-2.1.x
-
-server:
-  port: 8084
-
-eureka:
-  client:
-    filterOnlyUpInstances: false
-    registerWithEureka: true
-    serviceUrl:
-      defaultZone:   http://eureka.springcloud.cn/eureka/
-    reference: META-INF/reference.json # optional
-
-
-management.server.port: 8088
-
